+++ src/algoritmos/ordenacao/MergeSort.h
@@ -2,8 +2,8 @@
 #define MERGESORT_H
 #include <fstream>
 
-void mergeSort(int *vet, int esq, int dir, unsigned long long int *numComp, unsigned long long int *numReg);
+void mergeSort(int *vet, int esq, int dir, unsigned long *numComp, unsigned long *numReg);
 
-void merge(int *vet, int esq, int med, int dir, unsigned long long int *numComp, unsigned long long int *numReg);
+void merge(int *vet, int esq, int med, int dir, unsigned long *numComp, unsigned long *numReg);
 
 #endif // MERGESORT_H
