+++ src/algoritmos/ordenacao/quickSortInsertion.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 
-void insertionQuickSort(int *a, int inicio, int fim, int m, unsigned long long int *numComparacoes, unsigned long long int *numTrocas)
+void insertionQuickSort(int *a, int inicio, int fim, int m, unsigned long *numComparacoes, unsigned long *numTrocas)
 {
 
 		if(fim - inicio < m)
